Implements HeaderWithLogin and HeaderWithoutLogin
