+++ gradle.properties
@@ -1,13 +1,8 @@
 #CHROME or FIREFOX
 browser=CHROME
 
-#features
-featureFolders =
-cucumberOptions =
-
 #Web Sites
-baseurl = http://www.phptravels.net/admin
-smartHouseBaseUrl = http://localhost:4200/home
+baseurl = http://localhost:4200/home
 
 #Waits
 implicitTimeWait=15
