+++ src/main/java/org/umssdiplo/automationv01/core/managepage/HeaderWithoutLogin.java
@@ -1,0 +1,16 @@
+package org.umssdiplo.automationv01.core.managepage;
+
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.umssdiplo.automationv01.core.utils.CommonEvents;
+
+public class HeaderWithoutLogin extends BasePage {
+
+    @FindBy(name = "reports_button")
+    private WebElement reportsTab;
+
+    public SHReport clickReportTab(){
+        CommonEvents.clickButton(reportsTab);
+        return new SHReport();
+    }
+}
