+++ src/main/java/org/umssdiplo/automationv01/core/utils/LoadPage.java
@@ -1,12 +1,16 @@
 package org.umssdiplo.automationv01.core.utils;
 
 import org.umssdiplo.automationv01.core.customwebdriver.ManageDriver;
-import org.umssdiplo.automationv01.core.managepage.Login.Login;
+import org.umssdiplo.automationv01.core.managepage.HeaderWithLogin;
+import org.umssdiplo.automationv01.core.managepage.HeaderWithoutLogin;
 
 public final class LoadPage {
-    public static Login loginPage() {
-        ManageDriver.getInstance().getWebDriver()
-                .navigate().to(PropertyAccessor.getInstance().getBaseUrl());
-        return new Login();
+    public static HeaderWithLogin headerPageWithLogin() {
+        ManageDriver.getInstance().getWebDriver().navigate().to(PropertyAccessor.getInstance().getBaseUrl());
+        return new HeaderWithLogin();
+    }
+    public static HeaderWithoutLogin headerWithoutLogin() {
+        ManageDriver.getInstance().getWebDriver().navigate().to(PropertyAccessor.getInstance().getBaseUrl());
+        return new HeaderWithoutLogin();
     }
 }
